+++ lib-garble-rs/src/garble/new_garbling_scheme/wire_value.rs
@@ -1,0 +1,26 @@
+use serde::{Deserialize, Serialize};
+
+/// Represent a Wire's value, so essentially ON/OFF <=> a boolean
+#[repr(transparent)]
+#[derive(PartialEq, Debug, Serialize, Deserialize, Default, Clone)]
+pub(crate) struct WireValue {
+    pub(crate) value: bool,
+}
+
+impl PartialEq<bool> for WireValue {
+    fn eq(&self, other: &bool) -> bool {
+        &self.value == other
+    }
+}
+
+impl PartialEq<bool> for &WireValue {
+    fn eq(&self, other: &bool) -> bool {
+        &self.value == other
+    }
+}
+
+impl From<bool> for WireValue {
+    fn from(value: bool) -> Self {
+        Self { value }
+    }
+}
